requirements.txt
aiokafka==0.7.1
Faker==8.12.1
kafka-python==2.0.2
python-dateutil==2.8.2
six==1.16.0
text-unidecode==1.3
